+++ packages/components/Button/src/Button.types.ts
@@ -6,7 +6,7 @@ import { IPressableProps } from '@fluentui-react-native/pressable';
 import { FontTokens, IForegroundColorTokens, IBackgroundColorTokens, IBorderTokens } from '@fluentui-react-native/tokens';
 import { IFocusable, IPressableState } from '@fluentui-react-native/interactive-hooks';
 import type { IViewProps } from '@fluentui-react-native/adapters';
-import { IconProps , IconSourcesType} from '@fluentui-react-native/icon';
+import { IconProps, IconSourcesType } from '@fluentui-react-native/icon';
 
 export const buttonName = 'Button';
 
@@ -30,7 +30,7 @@ export interface IButtonInfo extends IPressableState {
   /**
    * End icon.
    */
-   endIcon?: boolean;
+  endIcon?: boolean;
 }
 
 /*