+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Tue, 14 Sep 2021 19:31:04 GMT and should not be manually modified.
+This log was last generated on Tue, 21 Sep 2021 16:32:12 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.43.3
+
+Tue, 21 Sep 2021 16:32:12 GMT
+
+### Patches
+
+- Added Experimental MenuButton (v.kozlova13@gmail.com)
 
 ## 0.43.2